+++ noodles-cram/src/record/convert.rs
@@ -121,7 +121,7 @@ impl Record {
         }
 
         if !self.bam_flags().is_unmapped() {
-            let cigar = resolve_features(self.features(), self.read_length());
+            let cigar = resolve_features(self.features(), self.read_length())?;
             builder = builder.set_cigar(cigar);
         }